fix from "shattered" (see hxc forum : http://torlus.com/floppy/forum/viewtopic.php?f=19&t=3457) :
"It's working now -- tested on a real C64 and 1451 drive, using "index" writing mode.

SCP software expects C64 images to be 96tpi, so I'm doubling every track (this is interpreted as image with half tracks, which probably helps -- my drive is 96tpi). 
Setting number of heads is apparently not critical -- SCP software expects C64 images to be single sided anyway -- but length of second gap is critical."